pyproject.toml
[project]
name = "rhea-sdk"
version = "0.1.4"
description = "Asynchronous SDK for interacting with Rhea Finance DEX"
authors = [{ name = "Maksim_Antsiferov", email = "antsiferov.ms@gmail.com" }]
license = { file = "LICENSE" }
keywords = ["python", "near", "rhea", "dex"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "asyncio>=3.4.3",
    "base58>=2.1.1",
    "pynacl>=1.5.0",
]

[project.urls]
homepage = "https://github.com/MaximAntsiferov/rhea-sdk"
documentation = "https://github.com/MaximAntsiferov/rhea-sdk"
repository = "https://github.com/MaximAntsiferov/rhea-sdk"
changelog = "https://github.com/MaximAntsiferov/rhea-sdk/blob/main/CHANGELOG.md"
twitter = "https://twitter.com/MaksimA30"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true